+++ CHANGELOG.md
@@ -19,6 +19,8 @@
  - Added trigger uk.dropped event when files were dropped to upload add-on
  - Added trigger uk.offcanvas.hide when offcanvas hides
  - Added allow multiple connect container to switcher component
+ - Added switch accross multiple connected switcher containers
+ - Added switch active content from within connected switcher containers via `data-uk-switcher-item` attribute
  - Updated active datepicker closes on window resize
  - Refactored sticky add-on and fixed z-index and box sizing
  - Refactored responsive tab behavior
@@ -33,6 +35,7 @@
  - Fixed animation reverse timing function
  - Removed Google maps automatic URL detection
  - Removed search close button and loading spinner
+ - Removed Promise polyfill
 
 ### 2.8.0 (June 19, 2014)